tanya and alina, displaying infromation from the DB on profile page, to each entry on list
