+++ comps/Navbar.js
@@ -8,7 +8,7 @@ const Navbar = () => {
       <Link href='/about'>פרופילים באינסטגרם</Link>
       <Link href='/'>עמוד הבית</Link>
       <Link href='/profileDisplay'> פרופיל מתחזה</Link>
-      <Link href='imposters'> רשימת המתחזים</Link>
+      <Link href='/imposters'> רשימת המתחזים</Link>
       <div className='logo'>
         <Image src='/logo.png' alt='site logo' width={128} height={77} />
       </div>
