+++ comps/ShowList.js
@@ -7,6 +7,7 @@ import ListItemAvatar from '@mui/material/ListItemAvatar';
 import Avatar from '@mui/material/Avatar';
 import Typography from '@mui/material/Typography';
 import { useGlobalContext } from '../public/context';
+import Link from 'next/link'
 
 export default function ShowList({ imposters, isPending }) {
   //todo: get from db file
@@ -23,26 +24,28 @@ export default function ShowList({ imposters, isPending }) {
   return (
     <List sx={{ width: '100%', maxWidth: 360, bgcolor: 'background.paper' }}>
       {imposters.map((imposter) => (
-        <ListItem key={imposter.id} alignItems='flex-start'>
-          <ListItemAvatar>
-            <Avatar src={imposter.img_url} />
-          </ListItemAvatar>
-          <ListItemText
-            primary={(imposter.first_name, imposter.last_name)}
-            secondary={
-              <React.Fragment>
-                <Typography
-                  sx={{ display: 'inline' }}
-                  component='span'
-                  variant='body2'
-                  color='text.primary'
-                >
-                  {imposter.timestamp}
-                </Typography>
-              </React.Fragment>
-            }
-          />
-        </ListItem>
+        <Link href={`profileDisplay/${imposter.id}`} key={imposter.id}>
+          <ListItem alignItems='flex-start'>
+            <ListItemAvatar>
+              <Avatar src={imposter.img_url} />
+            </ListItemAvatar>
+            <ListItemText
+              primary={(imposter.first_name, imposter.last_name)}
+              secondary={
+                <React.Fragment>
+                  <Typography
+                    sx={{ display: 'inline' }}
+                    component='span'
+                    variant='body2'
+                    color='text.primary'
+                  >
+                    {imposter.timestamp}
+                  </Typography>
+                </React.Fragment>
+              }
+            />
+          </ListItem>
+        </Link>
         // <Divider variant='inset' component='li' />
       ))}
     </List>
